display cards componetized out as an organism
