+++ src/components/pages/PageKittens/PageKittens.jsx
@@ -1,14 +1,9 @@
 import React from 'react'
 import { kittens } from "../../../kittens";
 import TemplateDefault from "../../templates/TemplateDefault/TemplateDefault";
+import ProfileCard from "../../organisms/ProfileCard/ProfileCard";
 
 function PageKittens() {
-
-  const makeSound = (e) => {
-    e.preventDefault();
-    console.log('meow test');
-    window.alert('meow , meow!')
-  }
 
   return (
     <TemplateDefault>
@@ -26,29 +21,7 @@ function PageKittens() {
 
         {
           kittens.map((animal) => (
-            <div
-              className="col-12 col-sm-6 col-md-4 col-lg-3 py-2"
-              key={animal.title}
-            >
-              <div className="card">
-                <img
-                  className="card-img-top"
-                  src={animal.image}
-                  alt="Card image cap"
-                />
-                <div className="card-body">
-                  <h5 className="card-title">{animal.title}</h5>
-                  <p className="card-text">{animal.description}</p>
-                  <button
-                    className="btn btn-primary"
-                    type="button"
-                    onClick={makeSound}
-                  >
-                    meow
-                  </button>
-                </div>
-              </div>
-            </div>
+            <ProfileCard animal={animal} />
           ))
         }
