+++ src/components/templates/TemplateDefault/TemplateDefault.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Link } from "react-router-dom";
 import Header from "../../organisms/Header/Header";
 import Footer from "../../organisms/Footer/Footer";
